[FLINK-21505] Extract SnapshotExecutionType from SnapshotStrategyRunner
