+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/tasks/TaskCheckpointingBehaviourTest.java
@@ -63,6 +63,7 @@ import org.apache.flink.runtime.state.DefaultOperatorStateBackendBuilder;
 import org.apache.flink.runtime.state.OperatorStateBackend;
 import org.apache.flink.runtime.state.OperatorStateCheckpointOutputStream;
 import org.apache.flink.runtime.state.OperatorStateHandle;
+import org.apache.flink.runtime.state.SnapshotExecutionType;
 import org.apache.flink.runtime.state.SnapshotResources;
 import org.apache.flink.runtime.state.SnapshotResult;
 import org.apache.flink.runtime.state.SnapshotStrategy;
@@ -372,7 +373,7 @@ public class TaskCheckpointingBehaviourTest extends TestLogger {
                                     "Failing strategy",
                                     FAILING_STRATEGY,
                                     registryForStateBackend,
-                                    SnapshotStrategyRunner.ExecutionType.ASYNCHRONOUS));
+                                    SnapshotExecutionType.ASYNCHRONOUS));
                 }
             }.build();
         }
